// src/typescript/data/web/run.d.ts
/**
 * 
 * @export
 * @interface Run
 */
export interface Run {
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof Run
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof Run
     */
    iteration: number;
    /**
     * 
     * @type {number}
     * @memberof Run
     */
    runtimeSecond: number;
}